+++ src/interface/alert.tsx
@@ -1,0 +1,4 @@
+export interface IAlert {
+  setVisibleAlert: React.Dispatch<boolean>;
+  value: string;
+}
