+++ src/interface/index.tsx
@@ -1,0 +1,18 @@
+import { IUserState, IUserAction } from "./user";
+import { IMessageBallon, IMessageState, IMessageAction } from "./message";
+import { IResponsiveSizeState, IResponsiveSizeAction } from "./responsiveSize";
+import { IInputEmoticon, IEmoticonPopover } from "./message";
+import { IAlert } from "./alert";
+
+export type {
+  IUserState,
+  IUserAction,
+  IInputEmoticon,
+  IEmoticonPopover,
+  IAlert,
+  IMessageBallon,
+  IMessageState,
+  IMessageAction,
+  IResponsiveSizeState,
+  IResponsiveSizeAction,
+};
