+++ src/interface/message.tsx
@@ -1,0 +1,30 @@
+import { IUserType } from "./user";
+
+interface IMessageType {
+  id: number;
+  text: string;
+  user: IUserType;
+  time: string;
+}
+
+export interface IMessageBallon {
+  message: IMessageType;
+}
+
+export interface IMessageState {
+  messages: IMessageType[];
+  dispatch: React.Dispatch<IMessageAction>;
+}
+
+export type IMessageAction = {
+  type: "ADD_MESSAGE";
+  message: IMessageType;
+};
+
+export interface IInputEmoticon {
+  addEmoticonMessage: (emo: string) => void;
+}
+
+export interface IEmoticonPopover extends IInputEmoticon {
+  handlePopover: () => void;
+}
