+++ src/interface/responsiveSize.tsx
@@ -1,0 +1,9 @@
+export interface IResponsiveSizeAction {
+  type: "RESIZE";
+  isMobile: boolean;
+}
+
+export interface IResponsiveSizeState {
+  isMobile: boolean;
+  dispatch: React.Dispatch<IResponsiveSizeAction>;
+}
